services/users/requirements.txt
Flask==1.0.2
psycopg2==2.7.6.1
Flask-SQLAlchemy==2.3.2
Flask-testing==0.7.1
gunicorn==19.9.0
coverage==4.5.2
flake8==3.6.0
flask-debugtoolbar==0.10.1
flask-cors==3.0.7
flask-migrate==2.3.1